MiningUtils/Coinbase/requirements.txt
python-bitcoinrpc==1.0
requests>=2.25.0